Upload project
The project is divided into a server and client part.
A scene with the ability to register a user and log in has been added.
A game scene has been added.
The ability to exit the game scene to the login window has been added.
The ability to get additional gold by clicking on the mine has been added.
A level up button has been added.
All processes are synchronized with the database.